Change pruning distance parameter from double to int
Also, do not let islands overlap by nothru edges as this can generate islands,
which are almost identical (only a few unique normal edges, lots of shared nothru edges).
Almost identical islands make issue report harder to use.

+++ src/main/java/org/opentripplanner/standalone/config/buildconfig/IslandPruningConfig.java
@@ -11,7 +11,7 @@ public class IslandPruningConfig {
 
   public int pruningThresholdIslandWithoutStops;
   public int pruningThresholdIslandWithStops;
-  public double adaptivePruningDistance;
+  public int adaptivePruningDistance;
   public double adaptivePruningFactor;
 
   public static IslandPruningConfig fromConfig(NodeAdapter root) {
@@ -64,7 +64,7 @@ public class IslandPruningConfig {
         The distance after which disconnected sub graph is considered as real island in pruning heuristics.
         """
         )
-        .asDouble(250);
+        .asInt(250);
 
     islandPruning.adaptivePruningFactor =
       config
